chore(evm): move exece input/output to types crate
