+++ crates/evm/execution-types/src/lib.rs
@@ -8,8 +8,11 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-mod execution_outcome;
-pub use execution_outcome::*;
-
 mod chain;
 pub use chain::*;
+
+mod execute;
+pub use execute::*;
+
+mod execution_outcome;
+pub use execution_outcome::*;
