+++ crates/evm/src/either.rs
@@ -2,11 +2,9 @@
 
 use std::fmt::Display;
 
-use crate::execute::{
-    BatchExecutor, BlockExecutionInput, BlockExecutionOutput, BlockExecutorProvider, Executor,
-};
+use crate::execute::{BatchExecutor, BlockExecutorProvider, Executor};
 use reth_execution_errors::BlockExecutionError;
-use reth_execution_types::ExecutionOutcome;
+use reth_execution_types::{BlockExecutionInput, BlockExecutionOutput, ExecutionOutcome};
 use reth_primitives::{BlockNumber, BlockWithSenders, Receipt};
 use reth_prune_types::PruneModes;
 use reth_storage_errors::provider::ProviderError;
