+++ crates/evm/src/noop.rs
@@ -3,15 +3,13 @@
 use std::fmt::Display;
 
 use reth_execution_errors::BlockExecutionError;
-use reth_execution_types::ExecutionOutcome;
+use reth_execution_types::{BlockExecutionInput, BlockExecutionOutput, ExecutionOutcome};
 use reth_primitives::{BlockNumber, BlockWithSenders, Receipt};
 use reth_prune_types::PruneModes;
 use reth_storage_errors::provider::ProviderError;
 use revm_primitives::db::Database;
 
-use crate::execute::{
-    BatchExecutor, BlockExecutionInput, BlockExecutionOutput, BlockExecutorProvider, Executor,
-};
+use crate::execute::{BatchExecutor, BlockExecutorProvider, Executor};
 
 const UNAVAILABLE_FOR_NOOP: &str = "execution unavailable for noop";
